+++ Solutions/SailPointIdentityNow/ReleaseNotes.md
@@ -1,4 +1,5 @@
 | **Version** | **Date Modified (DD-MM-YYYY)** | **Change History** |
 |---|---|---|
+| 3.0.2       | 16-09-2026                     | **Poller Config** "name" updated    |
 | 3.0.1       | 14-09-2026                     | Added CCF **Data Connector** support with one  **Parser** for backward compatibility and new schema for SailPoint IdentityNow events; updated **Analytic Rules** to use the parser alias and new connector ID |
 | 3.0.0       | 28-08-2024                     | **Data Connector** instruction updated      |